Fix receipt + refund cap to render decimal quantities
Item quantity was int-cast in the data sent to the print service (and
again in the ESC/POS helper), so a 3.5-unit line printed as "3 x $10.80
= $37.80" — internally inconsistent and confusing for customers and
staff. Refund UI capped max refundable at floor(quantity) for the same
reason, blocking a full refund of fractional lines.

Cast quantity as float in the three JSON builders that feed the print
service, format it as a trimmed decimal in escpos_helpers (3.0 -> "3",
3.5 -> "3.5"), and switch the refund-qty input to float + step=0.01.

+++ app/controllers/ApiController.php
@@ -294,7 +294,7 @@ class ApiController {
                 'cashier'       => $transaction['username'] ?? '',
                 'items'         => array_map(fn($i) => [
                     'name'             => $i['product_name'],
-                    'quantity'         => (int)$i['quantity'],
+                    'quantity'         => (float)$i['quantity'],
                     'unit_price'       => (float)$i['unit_price'],
                     'line_total'       => (float)$i['line_total'],
                     'gst'              => (float)$i['gst'],
