+++ app/views/transactions/refund_receipt.php
@@ -74,7 +74,7 @@ var _refundReceiptData = <?= json_encode([
     'cashier'            => $refund['refunded_by_name'] ?? '',
     'items'              => array_map(fn($i) => [
         'name'       => $i['product_name'],
-        'quantity'   => (int)$i['quantity'],
+        'quantity'   => (float)$i['quantity'],
         'unit_price' => (float)$i['unit_price'],
         'line_total' => (float)$i['line_total'],
         'gst'        => (float)$i['gst'],
